backend/pymc/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
arviz==0.17.1
async-timeout==4.0.3
attrs==23.2.0
bayesian-optimization==1.4.3
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
cma==3.3.0
colorama==0.4.6
cons==0.4.6
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
db-dtypes==1.2.0
dill==0.3.8
etuples==0.3.9
fastprogress==1.0.3
filelock==3.14.0
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Session==0.6.0
fonttools==4.48.1
google-api-core==2.17.0
google-api-python-client==2.126.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==20.1.0
h5netcdf==1.3.0
h5py==3.11.0
htmlmin==0.1.12
httplib2==0.22.0
idna==3.6
ImageHash==4.3.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.6
llvmlite==0.41.1
logical-unification==0.4.6
MarkupSafe==2.1.5
matplotlib==3.8.2
miniKanren==1.0.3
multimethod==1.11
multipledispatch==1.0.0
multiprocess==0.70.14
nestd==0.3.1
networkx==3.2.1
nevergrad==1.0.1
numba==0.58.1
numpy==1.25.2
oauthlib==3.2.2
packaging==23.2
pandas==1.5.3
pandas-gbq==0.22.0
patsy==0.5.6
pfzy==0.3.4
phik==0.12.4
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydata-google-auth==1.8.2
pymc==5.12.0
pymc-marketing==0.4.2
pyparsing==3.1.1
pytensor==2.19.0
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.12.2
six==1.16.0
slack_sdk==3.27.1
statsmodels==0.14.1
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
toolz==0.12.1
tqdm==4.66.1
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
visions==0.7.5
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wordcloud==1.9.3
xarray==2024.3.0
xarray-einstats==0.7.0
ydata-profiling==4.6.4
zipp==3.17.0
celery==5.3.6
